+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - added plugins for docker and ecr
 - introduced "::before" and "::after" keywords for plugins
 - added deployment mode
+- support for /etc/default/<application>.switch.json config files, so we
+  can set mode and type application specific (is going to replace
+  /usr/local/bin/<application> -i TYPE)
 
 0.1.0 (2019-03-01)
 ==================